+++ PasterDream/src/main/java/com/pasterdream/pasterdreammod/item/ShadowBreathItem.java
@@ -18,7 +18,7 @@ public class ShadowBreathItem extends Item {
      * @param properties 物品属性
      */
     public ShadowBreathItem(Item.Properties properties) {
-        super(properties.stacksTo(64));
+        super(properties.stacksTo(1));
 }
 
     @Override